Separating base class
